GTK2/3 synchronization for Progressbars and Sliders.
This commit adds the barebones needed to achieve a synchronized look to
the GtkProgressBar and Scale/Slider widgets.

In particular, the progressbar is defaulted back to the classic "wide"
look default for 2/3, though animation for it is removed since it's not
currently clear how to achieve the same effect in both engines.