GtkNotebook 2/3 synchronization.
This commit synchronizes the look of the GTK2 notebook widget to GTK3
as much as possible.

This switches the "pixmap" engine to "murrine", and matches the tab
spacing and height between the two elements. There is a slight residual
highlight effect left behind, but this is minimized.

The removal of the pixmap engine solves a number of unsightly glitches
throughout GTK2 applications when labels and other elements are
superimposed.