Fix smudgy appearance of treeview expanders.
The treeview expanders looked very "smudgy" in the current theme due to
an imbalance of the grey colors around them. This commit removes the
imbalance and replaces it with clean transparencies and sharp lines,
greatly improving the overall appearance.

The expander icon size property is changed to match the new expander
icons in GTK2/3.