[Mint-X/Cinnamon] Fix modal dialogs button shifting when buttons are focused
The modal dialog buttons (class **modal-dialog-button**) had a left and right margins added when focused (**focus** and **focus:hover** pseudo-classes). I removed the margins so the buttons position is kept when switching focus.

The following animated image illustrates the problem (the **Tab** key is pressed to switch focus):

![ezgif-5-4efc2ff83e](https://user-images.githubusercontent.com/3822556/38509735-28355e0c-3bf9-11e8-963c-4a2c8417ca74.png)

The following animated image illustrates the fixed behavior:

![ezgif-5-a5acdc20f2](https://user-images.githubusercontent.com/3822556/38510398-56c49312-3bfb-11e8-88de-33729dedf26e.png)

**Note:** I opted by removing the margins instead of adding them to the **modal-dialog-button** selector so the alignment of the buttons in relation to the other elements of the dialog are somewhat symmetric.