Mint-X-Cinnamon: Fix color of selected window in the workspace switcher
The green we were using was slightly different so it wasn't being properly
replaced by the sed command in generate_themes. Tweaking the color a bit fixes
the issue.

This also turned a typo in generate_themes.py that was causing us to get red
highlights in places in the orange variation. Fix that up as well.