Add files via upload
This is a version of `metacity-theme-3.xml`; I have modified that file from the version of it to be found here:

https://github.com/smurphos/Window_Borders_Mint_19

That file (the one at https://github.com/smurphos/Window_Borders_Mint_19) was created by 'smurphos' from the version of the file that shipped with Mint 18.3. The modifications made by smurphos consist in making the window buttons (close, maximise . .) larger. *My* modifications consist in adding, to (only) the dark version the theme, some slightly shaded titlebars, the idea being to make windows more easily distinguishable from each other.