Style menu favorites buttons as squares (#157)
* Mint-Y: Correctly style menu favorites buttons

Styles from .menu-favorites-button were not being applied because
they are also .popup-menu-item and that class properties had a
higher CSS priority due to the class being nested. So remove nesting.

Also add some padding to the favorites buttons because it was very
small.

* Mint-X: Use square menu favorites buttons

It's not mathematically a square because human eye vertically
stretches images. This looks subjectively more like a square
than an actual square. There are tons of articles that talk
about this.