Cinnamon: Increase padding and center align text in applets
When the clock uses %n in its date format, it looks wrong because
the text is left-aligned. This fixes it.

The commit also increases the applet padding from 3px to 4px in
Mint-Y (it was already 4px in Mint-X). This is to detach the clock
and the menu a bit more from the edges of the screen and to increase
slightly the space between status icons to better match the spacing
between systray icons.