Reverted Colorize GTK2, back in place on lines 123-129
I did this backward change so this master branch does not differ too much from linuxmint/mint-themes. Going one step at a time. But i am still not sure this "Colorize GTK2" is the best thing to do. I think i had error message(s) when running this part. And also, it looks like the missing assets are similar to the ones that can be found in /assets/... This needs further investigation and tests...