Mint-Y: Add a dark variant to the light theme
Now that Mate properly supports GTK_THEME_VARIANT we add a dark version of
the theme. This allows apps that want to use the dark theme to use it.

Also adapt the metacity-3 theme to properly pull the colors from the gtk
theme so window borders will match properly in dark applications.