Mint-Y: Restyle the window decorations
Revamp both the metacity and gtk themes to use a new decoration style. This
is similar to the old style but has larger buttons and new icons. For the
gtk changes we now use css and themed icons instead of relying on all image
based assets. This makes changing things moving forward much easier and allows
for the removal of 200+ assets from the theme.