Mint-X-Gtk3: Port the gtk3 theme to sass
This is a large commit. It ports the theme to gtk3 and adds a dark variation
to support themes that allow using one. It removes a large number of colored
assets and draws them instead with css.

Support for the metacity1 and 2 variations is also removed. For now fringe
button support is also removed.

This probably needs some more testing and tweaking, especially under Mate and
Xfce.